requirements.txt
pandas 
numpy  
matplotlib
scikit-learn
python-binance
prophet
cryptography
python-dotenv
python-prophet